fix: move command parser into separate utils module.
